Update Buff
Buff 아이템 2개 한꺼번에먹으면 안되던버그 수정

+++ Client/Source/Project1/UI/SelectCharacterLevel/CharacterDescWidget.cpp
@@ -63,7 +63,7 @@ void UCharacterDescWidget::SetDesc(EPlayerJob PlayerJob, const TMap<EPlayerJob, 
 	String += FString::Printf(TEXT("MP\t : %d\n"), static_cast<int>(CharacterInfo->m_SP));
 	String += FString::Printf(TEXT("ATK\t : %d\n"), static_cast<int>(CharacterInfo->m_ATK));
 	String += FString::Printf(TEXT("DEF\t : %d\n"), static_cast<int>(CharacterInfo->m_DEF));
-	String += FString::Printf(TEXT("AttackSpeed\t : %d\n"), static_cast<int>(CharacterInfo->m_AttackSpeed));
+	String += FString::Printf(TEXT("AttackSpeed\t : %.2f\n"), CharacterInfo->m_AttackSpeed);
 	String += FString::Printf(TEXT("MoveSpeed\t : %d\n"), static_cast<int>(CharacterInfo->m_MoveSpeed));
 
 	FString ClassName = TEXT("");
